update mobile ui
