--- public/audio.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-  </head>
-  <body>
-    <audio
-      width="{300}"
-      height="{300}"
-      src="https://file-examples.com/storage/fe00d37cde6728af4966ebc/2017/11/file_example_MP3_700KB.mp3"
-      autoplay
-    ></audio>
-  </body>
-</html>
